fix: rename hook file
